pyproject.toml
[tool.poetry]
name = "social2-0"
version = "0.1.0"
description = ""
authors = ["daron035 <kamil249@mail.ru>"]
readme = "README.md"
packages = [{include = "social2"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
pillow = "^9.3.0"
djoser = "^2.1.0"
django-cors-headers = "^3.13.0"
django-stubs = "^1.13.0"
coverage = "^6.5.0"
mock = "^4.0.3"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"
channels = "^4.0.0"
gunicorn = "^20.1.0"
daphne = "^4.0.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "src.settings"
